requirements.txt
numpy
Flask
gunicorn
pandas
PyYAML
evidently
matplotlib
scikit-learn
joblib
black
pyunpack
requests
patool
dill
-e .